MAIL changed profile labels in the C codegen
There was a problem that would show up when using the C codegen and
newer gcc's.  We used to have DeclareProfileLabel and ProfileLabel in
the C codegen defined as

#define DeclareProfileLabel(l)			\
	void l() __attribute__ ((weak))

#define ProfileLabel(l)				\
	__asm__ __volatile__ (#l ":" : : )

DeclareProfileLabel caused the label to be a global symbol defined in
the text segment, where the actual definition got stuck into the
assembly code.  But, with newer gcc's (3.2.2, 3.3.2), for some reason
the symbol is not global, and hence is not available to the profiling
infrastructure.  So, I changed DeclareProfileLabel and ProfileLabel to

#define DeclareProfileLabel(l)			\
	void l() __attribute__ ((alias (#l "_internal")))

#define ProfileLabel(l)				\
	__asm__ __volatile__ (#l "_internal:" : : )

This does cause the label to be global with both newer and older
gcc's.