MAIL new cmcat
A new cmcat, which meets (or exceeds) the behavior of the old cmcat.
After building the import graph, it filters all files not necessary to
yield the exports of the input .cm file.  We avoid chasing imports to
CM anchored libraries (like "$/basis.cm"), but do track the
dependencies.  If you run cmcat with "-comments", then the output list
will include the dependencies in (SML-style) comments.

The new cmcat determines that:
  atoms/scheme.sig
  atoms/type.fun
  atoms/type.sig
  xml/sxml-tree.fun
are unnecessary for mlton/sources.cm.