MAIL -type-error {concise|full}
Added a switch that controls the display of type error messages.  With
-type-error concise, only the components that don't unify are shown.
With -type-error full, all of the type is shown.  In both cases, to
make it easier to visually spot the differences, brackets are placed
around the parts that aren't unifiable.

Consider this program

	fun f {w: int, x: int, y: real} = 13
	val _ = f {w = 13, x = 13.0, z = 14}
	fun f (x: int, y: real) = 13
	val _ = f (1, 2, 3)
	fun f (x: int, y: real) = 13
	val _ = f (1, 2)

With -type-error concise, the error messages are:

Error: z.sml 2.9: function applied to incorrect argument
   expects: {x: [int], [y]: _, ...}
   but got: {x: [real], [z]: _, ...}
   in: f {w = 13, x = 13.0, z = 14}
Error: z.sml 4.9: function applied to incorrect argument
   expects: [int * real]
   but got: [int * int * int]
   in: f (1, 2, 3)
Error: z.sml 6.9: function applied to incorrect argument
   expects: _ * [real]
   but got: _ * [int]
   in: f (1, 2)

With -type-error full, the error messages are:

Error: z.sml 2.9: function applied to incorrect argument
   expects: {w: int, x: [int], [y]: real}
   but got: {w: int, x: [real], [z]: int}
   in: f {w = 13, x = 13.0, z = 14}
Error: z.sml 4.9: function applied to incorrect argument
   expects: [int * real]
   but got: [int * int * int]
   in: f (1, 2, 3)
Error: z.sml 6.9: function applied to incorrect argument
   expects: int * [real]
   but got: int * [int]
   in: f (1, 2)

For now, -type-error concise is the default.  Unlike some of our
earlier ideas based on print-graph from Scheme, I chose to bracket the
parts of the types that are different, instead of those that are the
same.  I think that makes more sense since we care more about the
differences when there is a unification error.  I decided to use
brackets because {} and () are already taken and I want something that
matches.

I'd like people to try these out and send

* bug reports
* thoughts on which should be the default.  Do we even need both?
* thoughts on other improvements

Once we sort this out (soon), I'd like to make an experimental release
and announce it on MLton-user to start getting more feedback on the
front end (and all the other improvements).