MAIL improved type constructor names in type errors
Display type constructor names based on the environment currently in
scope.  Choose a shortest (in terms of dots) name for each type
constructor.  So, when the type name is in scope, the short name will
be used, as in

	structure S =
	   struct
	      datatype t = T
	      fun f T = ()
	      val _ = f 13
	   end

	Error: z.sml 5.15: function applied to incorrect argument
	   expects: [t]
	   but got: [int]
	   in: f 13

However, once we are outside the structure S, then the full name will
be used:

	structure S =
	   struct
	      datatype t = T
	      fun f T = ()
	   end
	val _ = S.f 13

	Error: z.sml 6.9: function applied to incorrect argument
	   expects: [S.t]
	   but got: [int]
	   in: S.f 13

Also, use a "?." to prefix type constructors that are not currently
accessible.  For example:

	datatype t = T
	fun f T = ()
	type t = unit
	val _ = f 13

	Error: z.sml 4.9: function applied to incorrect argument
	   expects: [?.t]
	   but got: [int]
	   in: f 13