MAIL non-exhaustive warning now displays an example pattern
The match compiler keeps track for each case of a pattern that leads
to that case, and feeds the information to the elaborator.  Then, when
the elaborator discovers a nonexhaustive match, it prints out the
corresponding pattern.

One slight annoyance was that by the time the match compiler sees
things, there is no distinction between char and word8.  So, I had to
add a little more information to the ILs to propagate the difference.

Added a few more examples to regression/nonexhaustive.sml.  Please
try out other examples that you think of.

Added an error message for real constants in patterns.