MAIL non-exhaustive warnings now output all omitted patterns
The implementation was simple enough.  The caller of the match
compiler now keeps a list of example patterns that fail.  The match
compiler itself uses an "or pattern" to capture all of the
constructors in a sum type that are not covered.  All of the example
patterns are displayed as one big or pattern.

I used or patterns to avoid the potential exponential blowup for cases
on tuples.  Also, I still only show one constant that is unhandled for
cases on constants.  Finally, clausal function definitions warnings
are now slightly messier, since the clauses are displayed as a tuple
instead of with spaces.

Overall, this seems like a nice improvement to me.  For example, for

	datatype z = A | B | C
	fun f x =
	   case x of
	      (A, B, C) => ()

we get the following warning

  missing patterns: (A, B, A | B) | (A, A | C, _) | (B | C, _, _)