MAIL where and sharing in signatures
Improved the implementation of flexible tycons, where clauses, and
sharing.  MLton now enforces the side conditions (t \not \in T of B)
on rule 63 and (t_i \not \in T of B) on rule 78.  This means that
signatures that in the past would have been accepted are now rejected.
I've added several regression tests to check for these failures.  I
had to update a few places in the basis library that incorrectly used
where or sharing, mostly attempting to redefine rigid tycons.

The implementation is much cleaner than before, because where clauses
now explicitly associate the type structure with the flexible tycon
being defined.  There was a slight trick to handle the mutual
recursion between the definition of flexible tycons and the TypeStr
functor, which I handled with an exn.