Added expert flag -dead-code {true|false}.
When making the world, checking of the basis library now includes
checking for nonexhaustive and redundant matches in addition to the
usual type checking.  To do this, I extended the meaning of -stop tc
so that it works with no input files.

Extended the meaning of -show-basis so that it works with an input
file.  With a file, MLton displays all the toplevel bindings of the
input file in addition to the basis.