MAIL change in handling of undetermined toplevel types
Changed the coalescing of topdecs so that a semicolon forces a
toplevel declaration to be treated as two separate topdecs.  This is
required by the restriction on page 14 of the Definition.  For
example, the following program now fails, because the overloading is
resolved for the double function without knowledge of its use.

fun double x = x + x;
val y = double 2.0

This was implemented by changing Ast.Program.t slightly to keep track
of where the semicolons are and change coalesce to only work between
them.

I also changed the handling of undetermined types at the top level.
It used to be that after the elaboration of each core declaration, we
would check for any undetermined types, and report an error if there
were any.  This was incorrect because the determination should be done
at the granularity of topdec, not core declaration.  Hence, the
following example, given on page 90 of the Definition, now elaborates.

structure A: sig val f: int -> int end =
   struct
      val f = (fn x => x) (fn x => x)
   end

Second, in order to prevent mistakenly rejecting programs where a core
dec has a free type variable, but it doesn't make it into the toplevel
environment, I decided to go for the more conservative approach of
instantiating the undetermined types with new type constructors.  So,
now the following program elaborates, as required by the Definition.

structure B : sig end =
struct
    val a = ref nil
end

What will still be rejected, and correctly I believe, is programs
where a type is determined across multiple topdecs.  For example, the
following will be rejected.

val x = ref [];
val _ = 1 :: !x

I am not sure if this is consistent with the Definition, but only
because the error message reports the error as being in the second
declaration (function applied to incorrect argument) rather than the
first (refusal to allow undetermined types to enter the basis).