Fixed a bug in constant overloading that's been there for a while. MLton would accept a program like
  val x = 0w0
  signature S = sig end
  val _ = x: Word8.word

and then report an internal IL type error.  The correct thing to do is
to reject the program, because the signature means that the "val x =
0w0" is its own topdec, and overloading must determine the type
without looking beyond that.

The bug was that although overloading resolved the type to be
Word.word, it left the type ununified with Word.word; hence the later
unification with Word8.word succeeded.