MAIL datatype declarations with free type variables
Allow datatype declarations with free type variables.  This is now
sound because we implement the rule that prevents a local datatype
from escaping its defining let.  It is implemented in the translation
from CoreML to XML, which is where datatypes are lifted to top level.
The translation adds any free type variables as extra arguments to the
type constructor and value constructors.  For example

  fun 'a f () =
     let
        ...
        datatype 'b t = T of 'a * 'b
        ...
        e : int t

moves the datatype to the top level as

  datatype ('a, 'b) t = T of 'a * 'b

and replaces "int t" with "('a, int) t".