Previously, when creating a StreamIO.instream from a reader and a non-empty buffer, StreamIO.filePosIn would raise IO.RandomAccessNotSupported because the base position of the buffer could not be determined (until an actual read from the underlying reader was performed, and the position could be verified).  There is no assurance that the buffer really came from the reader, or that it was the last items read from the reader.
Now, StreamIO.mkInstream'' takes a
{bufferContents: (bool * vector) option, ...},
where the boolean should be true exactly when the buffer is the last
items read from the buffer.  Hence, we properly set the base position
of the buffer when the boolean is true, the reader has the getPos
method, and the StreamIO functor has a translation from StreamIO.pos
to Position.int.  (This is always true for BinIO and TextIO, assuming
the underlying reader is to a regular file.)

StreamIO.mkInstream (exported to the user) always sets the boolean as
false.

ImperativeIO.getInstream always sets the boolean as true.

The following program demonstrates the difference:

      val f = BinIO.openIn "z"
      val _ = BinIO.input1 f
      val p = BinIO.StreamIO.filePosIn(BinIO.getInstream f)
      val _ = print (Position.toString p)
      val _ = print "\n"
      val _ = BinIO.closeIn f

Under the old code, this would raise IO.RandomAccessNotSupported.
Under the new code, this prints "1".