Added control elaborateOnly, which is set to true when both -stop tc and -warn-match false are set.
Added a simple hack to the elaborator to avoid doing any
defunctorization when elaborateOnly is set.  This saves a lot of time
because there is no need to take a snapshot of the environment at the
point of functor definition, and we don't need to generate any
declarations when applying functors.

With this hack, type checking all of mlton is back under 20 seconds.