MAIL major improvements to -show-basis
Pretty print signatures and functors when laying out a basis.

Wherever possible when laying out structures, use a signature
identifier to describe the structure, and add where clauses to define
the flexible types of the signature.

Please try this out, both for the basis library and for user programs
and send bug reports and suggestions for improvement.