MAIL added flag -sequence-unit {false|true}
If true, then in the sequence expression (e1; e2), it is a type
error if e1 is not of type unit.

I implemented it as an error rather than a warning, because it is
easily implemented by unifying the type of e1 with unit, which may
cause a type error.  To implement as a warning would require not
unifying, and delaying the check of e1's type until undetermined types
are reported, which seems to me like it could be more confusing that
immediately reporting the error.  For example, consider what should be
the inferred type of f in

val f =
   let
   in
      fn g => (g (); ())
   end

Forcing g to return type unit means that the type of f is "(unit ->
unit) -> unit".  If we didn't force g to return unit, then we wouldn't
know whether to issue a warning or not until we discover the type of
f.

Compiling the basis library with -sequence-unit true turned up a few
spurious errors, and even caught several bugs in cases where return
values from system calls were being ignored.

I also fixed the one spurious error in the MLton library.

There are a couple hundred such errors in the MLton sources proper.
I'll look into those at some point.