MAIL added flags -show-def-use, -warn-unused
Both of these deal with def-use information (as does
-show-basis-used).

-show-def-use <file>
	Causes MLton to output def-use information to file.

-warn-unused true
	Causes MLton to report unused identifiers.

The elaborator now keeps track of def-use information, which records
for each identifier (variable, signature, structure, functor, or type)
where each of its uses is.  Def-use information is a whole-program
property, and is tracked through structures and functor applications.
-show-def-use printsone line for each definition, and follows that
with an indented line for each use, showing the source position of
each use.  With an input program, -show-def-use and -warn-unused
report information for that program.  With no input, they report
information for the basis library.

-warn-unused reports all identifiers that have no uses, with a few
exceptions to cut down on spurious warnings.  First, identifers that
are still in scope are not reported.  Second, identifiers that are
defined in functor bodies of unused functors are not reported if they
are either exported by the functor or are passed as arguments to other
functors.  Third, types defined as datatypes are not reported, since
often the constructors are used but the type is not.

With these, I was able to go through the basis library and eliminate
all unused identifier warnings.  Most of the warnings were reasonable.
A few even showed a bug, where a variable should have been used but
wasn't.  For the few remaining cases, I have added a bogus use to
quell the warning.

Implementing this required some restructuring of the elaborator so
that I could track use information with constructors accessed via
their type structure.  For example, in

----------------------------------------
structure S =
   struct
      datatype t = A
   end
structure T =
   struct
      datatype z = datatype S.t
      val _ = A
   end
----------------------------------------

we would like to match up the definition of A with its use.  This
required augmenting type structures with use information.  This was a
bit messy, because the same code was used for type structures in
signatures as in structures.  I decided to split the implementation of
type structures into two, which on the whole was cleaner.

In adding -show-def-use, I cleaned up the code that calls the
elaborator and took Henry's suggestions to change -export-header,
-show-basis, and -show-basis-used to take a file name argument.  They
can be used together, or as part of a normal compilation.

I also cleaned up some hardwired stuff in compile.fun dealing with the
basis environment.  I added a new form of topdec, only to be used at
the end of the basis, that looks as follows

	_basis_done <longstrid>

The use looks like

	_basis_done MLtonFFI

Here, MLtonFFI identifies the structure that provides the primitives
that are needed in order to expand _export in the elaborator.
_basis_done also tells the elaborator that rebinding "=" is no longer
allowed.

There are still some more changes that will help cut down on spurious
warnings and will improve the output for -show-def-use.  But things
are working well enough that I'd like feedback.  Please try out
-show-def-use and -warn-unused on your programs and see what you
think.

Also, there are about 2300 unused warnings on MLton itself.  It would
be wortwhile to start plowing through those and understanding if they
are bugs, unused variables, or spurious warnings.