Introduced a new type, TyconMap.t, which describes paths to all the flexible tycons in an interface.  Each flexible tycon in the interface occurs in the tycon map exactly once.
When creating a tycon map, we make sure and choose a shortest path to
each flexible tycon.  To realize an interface, first copy it, then
create a tycon map, and then instantiate the flexible tycons.  This
will make pretty printing of interfaces better, since it uses
realization to choose names for the flexible tycons.

I also plan to use tycon maps to speed up some other aspects of module
elaboration.