MAIL Added Int31 structure
Most operations are implemented by converting to Int32.int and using
the corresponding operation there.

Added a new primitive "int31" tycon to the compiler, and cleaned up
the compiler internals a bit so that adding more int types should be
very easy.  The only int31 primitives that the compiler knows about
are the conversions to and from int32.  These conversions are replaced
by no-ops in ssa-to-rssa, which also replaces the int31 type with
int32, so that the codegens don't see anything new.

The reason for propagating the int31 type so far back is that soon the
representation pass will take advantage of the int31 type and will use
only 31 bits to represent it, which will allow sum types like "A of
Int31.int | B of Int31.int" to be represented more efficiently.

I also plan to add all the other int sizes between 1 and 64 soon.