Fixed SSA optimizer bug that could cause the error "<type> has no tyconInfo property".  The problem was in remove-unused, which removed a datatype declaration if none of its constructors were ever used to construct a value.  Unfortunately, the datatype tycon could still be used in types.  The SSA type checker doesn't check that all tycons are declared (or primitive) and so didn't catch the problem.  Usually, I guess the shrinker would get rid of stuff before a problem happened. But in one case, a self compile with a stub for IntInf.<< that raises Fail, the undeclared tycon made it all the way to the useless pass, which then died with the "no tyconInfo property" message.
I fixed the bug by changing remove-unused to keep all datatypes
around.  In the case where none of the constructors is useful, it
creates a single dummy constructor.

It would also be nice to fix the SSA type checker to check that all
tycons are declared.  I've added it to my todo list, but with a low
priority.