MAIL reorganized the handling of primitives in SsaToRssa
Now, for each codegen, there is an explicit statement of the
primitives that ssaToRssa expects the codegen to implement.  There is
also a specification of the primitives that are implemented by
functions in the runtime library.  Using these, the translation of
primitives is greatly simplified.  Either the primitive is handled by
ssaToRssa directly, or it is implemented by the codegen, and if not,
then it must be implemented by the runtime library or an error is
reported.

This should make it a lot easies to add a new codegen, which will
surely implement a different subset of the primitives than either of
our current codegens.

Changed a lot of int and word primitives in the basis library from
_import to _prim, with the decision on whether to call the runtime
library now delayed until ssaToRssa.  This exposes the prims to the
simplifier.  One reason for doing this is to make sure that arithmetic
on all the new int and word sizes can be simplified at compile time.

Tweaked the RSSA IL to record as part of the program itself whether or
not the program handles signals.  It used to be kept by leaving around
the MLton_installSignalHandler primitive.