MAIL better blocking/unblocking of signals in runtime
This adds
        /* signalsBlocked is the set of signals which the mutator has blocked;
	 * it is used in blockSignals and unblockSignals when blocking and
	 * restoring signals.
	 */
	sigset_t signalsBlocked;
to struct GC_state.
The sigprocmask in blockSignals uses signalsBlocked as the oldset; the
sigprocmask in unblockSignals uses signalsBlocked with SETMASK.

This gives the right behavior for enter/leave; no SML code runs
between these two functions.

It may not be the right behavior for GC_{start,finish}Handler.  The
problem is that between GC_startHandler and GC_finishHandler, the SML
signal handler is run.  If the handler calls
MLton.Signal.Mask.setBlocked, then this change will be lost when
GC_finishHandler is invoked.  Unfortunately, I suspect that blocking
signals is one thing you might want to do in the signal handler.