MAIL critical sections during thread switch
Added an implicit canHandle-- to GC_switchToThread.  This meant
adjusting a lot of the atomicBegin/End calls in thread.sml and
cont.sml.  I've annotated the code with the expected atomic state;
(technically, the least value of the atomic state; everything should
be fine if the calls are nested within an outer critical section; upon
returning to user code, the atomic state should be the same as upon
entry).

Rather than having two separate switchTo calls, I just put an extra
atomicBegin when switching to a New thread; this prevents switching
threads at the implicit canHandle-- in GC_switchToThread; once the new
thread is copied and ready to run, we atomicEnd.

While I was at it, I modified the ssa-to-rssa translation to force a
collection when atomicEnd drops the atomic state to 0 and a signal is
pending.