MAIL replaced type system in Rssa and Machine
The new type language, RepType (see atoms/rep-type.sig), is aimed
expressing bit-level control over layout and associated packing of
data representations.  There are singleton types that denote
constants, other atomic types for things like integers and reals, and
arbitrary sum types and sequence (tuple) types.  The big change to the
type system is that type checking is now based on subtyping, not type
equality.  So, for example, the singleton type 0xFFFFEEBB whose only
inhabitant is the eponymous constant is a subtype of the type Word32.

The type system makes lots of things cleaner than they used to be.
For example, we used have a magical IntInf type and coercions between
it and word vectors or integers.  Now, we can express IntInf as a sum
type of a pointer to a word vector and a sequence type that requires
the low bit to be one and the high 31 bits as an integer.  From the
code,

      val intInf: t =
	 sum (Vector.new2
	      (wordVector,
	       seq (Vector.new2
		    (constant (WordX.fromIntInf
			       (1, WordSize.fromBits (Bits.fromInt 1))),
		     int (IntSize.I (Bits.fromInt 31))))))

Now, the subtyping rules mean that we don't need anything special to
handle WordVector_toIntInf, since wordVector is a subtype of intInf.

Similarly, the special purpose EnumPointers type and rules that we
used to have now falls out as a sum type and the usual subtyping
rules.

Calls to C functions are now type checked in all ILs (they used to be
checked nowhere).  In order to do this, I needed to change the type
language for C functions to express their arguments and results as
RepTypes.  This caused some reorganization of code, moving stuff from
backend/ to atoms/, since C functions (and now hence RepTypes) are
used from the first IL.  One nice effect of all this is that the types
of C functions are now more clear.  For example, GC_copyThread now
takes gcState * thread instead of word * pointer.

Primapps are also now type checked more than they used to be.  See the
typeCheck function in atoms/prim.fun.  Most of the time, this does the
obvious thing (e.g. Int_add has the type int * int -> int).  However,
typeCheck is clever when performing word operations (andb, lshift,
orb, rshift,...) on types where some part of the argument is known
(because it is a singleton type).  This is to properly type check when
multiple objects are packed into a word and word operations are used
to construct/destruct the word.  As a side benefit, these rules check
other things that we used to not check, like the right shift that we
do to the header word of a heap object in order to do a case on the
tag.

There are a couple of drawbacks to the new type system.  First, type
checking of the Machine program can be rather slow.  It's slow enough
that I've turned it off for now.  I'm pretty sure that the speed
problems are due to the newness of the code and will be pretty easy to
fix.  A more serious problem is that the implementation of subtyping
is messy and is certainly incomplete, meaning that there are times
when it will say that t1 is not a subtype of t2 even though the set of
values denoted by a type t1 is a subset of the set of values denoted
by t2.  The type checker works for everything that SsaToRssa is
producing now, but it would be nice to make it complete (but not too
slow), or at least have a clearer understanding of its boundaries.

The plan going forward is to fix the speed problems and then to start
experimenting with packed tuple and datatype representations.  Now
that all the infrastructure is in place and the representation pass
has been refactored and well-isolated, it should be pretty easy to do
so.

Now for some minor points.

Simplified case statements in Rssa and Machine so there are only
cases on words.  Also, the exhaustivity check is now a simple call to
RepType.isSubtype.

Updated the copyright notice to read 2004 whenever I saw an old
date.  Please do the same in files whenever you modify them and they
still have an old date.

Cleaned up atoms/prim.fun so that prims really are just represented
by their name and there are no more linear lookups in the list of all
prims.

Simplified CTypes (c-type.fun) so that they don't depend on the integer,
real, or word size used for SML.  Now, CType is a simlpe datatype.

  datatype t = Pointer | Real32 | Real64 | Word8 | Word16 | Word32 | Word64

I eliminated the distinction between integers and words, since there
isn't much of a difference from the codegen's point of view.  It did
require me to make a few changes to c-chunk.h of the C codegen to
ensure that macros used to implement integer primtives treated
arguments as the right type.

Created the structures Bits, Bytes, and Words to help keep the units
straight when talking about memory amounts.

The type preThread is replaced by thread in the defunctorizer.  This
is so that calls to C functions can type check, as they only know
about threads.

Reworked the CM files with appropriate #ifdefs so that MLton and
SML/NJ use the same files.  So, cmcat is no longer needed to develop
with MLton.  This means that order now matters in the CM files.