Use the Pointer_get* primitives instead of importing C functions to get and set characters in C strings.
It would be nice to do similar things for lots of other imported C
functions in the basis library implementation that are just getting
and setting from C data structures.  The one thing we are missing is a
way to express in SML the offsets of fields in C structs.