MAIL changed typing for C functions
Another overhaul of types for C functions.  Now, CFunction.t is
polymorphic in the type used for its arguments and result.  Each IL
fills in the type needed for its type system.  The types are
translated by each pass, ending up with representation types in the
Rssa and Machine.

Because C functions are treated as primitives in XML and SSA,
primitives are now also polymorphic in the type.

Improved type checking of primapps.

Removed primitive pointer tycon (for C pointers) and associated types.
The basis library uses Word32.word.