MAIL better blocking/unblocking of signals in runtime and basis
Keep gcState.signalsBlocked in sync with the user program.
To do so, we keep an ML side representation of the currently blocked
signals.  Upon a MLton.Signal.Mask.{block,unblock,setBlocked}, we
compute the new set of blocked signals and call sigprocmask with
SIG_SETMASK.  (Note, this eliminates PosixPrimitive.Signal.how.)

On the runtime side, the sigprocmask in unblockSignals uses
signalsBlocked with SIG_SETMASK.

Two things:
1) there is a race condition in calling sigprocmask and updating the
     ML side representation; this will go away with
     Posix.Error.sysCall.
2) MLton.Signal.Mask.getBlocked (and the ML side representation) do
     not reflect the blocking of signals during the execution of a
     signal handler.  This is easily changed, but it seems more
     natural this way.