All the regressions and benchmarks now work with -representation packed.
Changed how weird-size integers are embedded in the normal size
integers, e.g. how Int25 is embedded in Int32.  We used to represent
the Int25 as an Int32.  But this didn't work as well with packing,
which would like the Int25 to be 25 bits.  So, I changed the
representation to be [Int25, 0x0:7].  This makes coercions between
Int25 and Int32, which used to be free, harder.  The coercions are now
implemented in the basis library EmbedInt functor.

This made the packed representation easy.  If we ever find it's too
slow, we could go back to the old way, at the expense of some
additional complexity in the SsaToRssa conversion, which has to figure
out when to mask out the high bits.