MAIL restructure of signal handling
This checkin moves the entire process of blocking and unblocking of
signals during a signal handler to the ML side.  Signals are only
blocked while extracting the set of handlers from the
gcState.pendingSignals sigset.  No user code is run while signals are
blocked.

The enter/leave functions of gc.c are also modified to treat runtime
functions as running in a critical section (i.e., with canHandle++ in
enter and canHandle-- in leave); this prevents limit from being
modified while in the runtime.  If a signal is caught on the C side
while in the GC, the limit will be reset to 0 during leave.