MAIL Added PosixError.SysCall structure.
There is a little bit of an ordering problem; in that SysCall.syscall
wants access to MLton.Signal.Mask to block and unblock signals when in
a critical section; on the other hand, MLton.Signal needs a lot of
infrastructure to be implemented.  To handle it, PosixError.SysCall
has a
  val blocker: (unit -> (unit -> unit)) ref
that MLton.Signal.Mask updates with the correct function for blocking
and unblocking.