MAIL thread tweaks
In gc.c, abstracted out functions atomic{Begin,End} and
maybeSwitchToHandler.  Eliminated the distinction between startHandler
and GC_startHandler, and changed the call to GC_startHandler from
thread.sml so that it is within a critical section, which is what
GC_startHandler now expects.

Fixed a type warning in the prototype for Posix_Signal_resetPending.