MAIL eliminated the distinction between integers and words
Eliminated the distinction between integers and words from all ILs.
Now, there are only words.  Integers are replaced by words immediately
after front end type checking.  There are no longer any integer
primitives; instead, there are signed and unsigned versions of word
primitives (when a distinction needs to be made).  This simplified
constant folding, as well as makes all the ILs and optimizer passes
slightly simpler because there's one less thing to worry about.  It
also makes the codegens simpler, because they have one less kind of
constant to worry about (there was some pretty messy code in the
x86-codegen for integer constants that is now gone).

The names of many word and integer primitives in the basis library
primitive.sml has changed.  Likewise, so have the names of these
primitives in the C codegen and runtime.

Moved the code that checks whether a codegen implements a primitive
from SsaToRssa into each of the codegens.  The backend (and SsaToRssa)
now takes "cogegenImplementsPrim" as a parameter.  This let me clean
up x86-mlton.fun in the x86 codegen, because the code for testing
whether a prim is implemented is separated from the code for
implementing the prim.

Fixed a bug in the SSA shrinker that could cause a label to be called
with the wrong number of arguments.  The problem was in the code that
tried to simplify a case expression if all branches went to the same
label.  It forgot to check that the label took no arguments.

Improved the implementation of weird-size integers, implementing the
coercion from big to small using a single comparison.