Fixed a bug with sharing constraints in signatures that would sometimes mistakenly treat two structures as identical when they shouldn't have been.  This would cause some programs to be mistakenly rejected.
The problem was that when sharing structures, two structures would be
made identical too often, in particular when they agreed on all the
tycons defined in them.  That is not a sufficient reason to make them
identical, though, since they may disagree on flexible tycons defined
elsewhere but used inside them.  The fix was to be more conservative
as to when interface structures are made identical, only allowing it
for those that correspond to toplevel signatures, which we know have
no flexible defined elsewhere.

Added a test to regression/sharing.sml to check for this.

I first tried turning off this optimization entirely, but that slows
down type checking a self compile by about a factor of two.  This
happens because the MLton sources make extensive use of structure
sharing.  Fortunately, it is almost always on structures with
top-level signatures, so the slightly less agressive optimization is
sufficient to recover the time.