MAIL -profile count
Added to the profiling infrastructure the ability to count function
calls and case branches.  This is enabled with the new flag

	-profile count

There is also an expert flag that controls whether branches are
profiled.

	-profile-branch {true|false}

This is implemented as with the rest of the profiling stuff by
inserting ProfileEnter and ProfileLeave annotations in the front end,
and letting the profile pass at the very end of the backend turn these
annotations into the appropriate calls.  -profile count is very much
like -profile alloc.  It inserts a call to GC_profileInc just after
each ProfileEnter.  One difference is that because I want to record
every count and the simplifier is allowed to simplify ProfileEnter
immediately followed by ProfileLeave, with -profile count the front
end also inserts a call to touch just after every ProfileEnter.  This
is sufficient to ensure that the annotations are never simplified
away.  Of course, it can affect program performance, but getting
accurate counts is more important.

One nice use of -profile count is as a code coverage tool.  With it,
you can run your programs on lots of different inputs, feed all the
mlmon.out files to mlprof, and see which branches were never taken and
which functions were never called.