MAIL improved analysis for ref flattening
After looking at some more examples, I decided that the local analysis
that I had implemented was too brittle and didn't get enough cases.
The main problem was the syntactic check to see if a ref that was
extracted from a tuple was only every assigned to or dereferenced.
There were many situations where the ref would be passed as an
argument, and so would not be flattened, despite the fact that the
place it was passed would only assign or dereference the ref.

So, I decided to switch to a less syntactic and more data-flow
analysis based approach.  This checkin implements the new analysis.

The idea is to compute for each occurrence of a ref type in the program
whether or not that ref can be represented as an offset of some object
(conapp or tuple).  As before, a unification-based whole-program with
deep abstract values makes sure the analysis is consistent.

The only syntactic part of the analysis that remains is the part that
checks that a variable bound to a value constructed by Ref_ref only
occurs once and that that occurrence is in the construction of an
object.  If so, and it is consistent with the rest of the flattening
annotations, we flatten that ref.  That's it.

I am more worried about space safety with this analysis, since it lets
a ref's containing object be kept alive quite a bit longer than the
old analysis.  This will be worth revisiting at some point.

I looked through the results of this analysis on the benchmarks and
some other tests, and it seems to get what I want.  Here's a few nice
examples.

As before, it gets some of the refs in the IO stack.

It gets singly-linked lists and should get Dan Wang's mutable cons
example.

For the barnes-hut benchmark, it flattens all four refs in the main
Space.body record type.

For boyer, there are two data structures with refs and both are
flattened.

For raytrace, it flattens the ref in the second field of the OObj
constructor.

For tsp, it flattens the prev and next fields of the ND constructor in
the tree datatype.

Feel free to try out the new analysis (via -diag-pass refFlatten) on
your code and see what happens.